Create SCCM Computer Name
this script when used with the variable OSDComputerName will name a machine using  a predefined 4 letter prefix and the serial number of the device obtained from the Bios.  It also determines the type (laptop, workstation, or MS Surface) based on the chassis type and whether or not it has a battery.  The surface also looks for the product name to include the word Surface.